pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cornerstone-forge"
version = "0.1.0"
description = "PhotonForge PDK for the Cornerstone (University of Southampton) silicon photonics foundry"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Flexcompute Inc." }]
keywords = ["photonics", "PDK", "Cornerstone", "PhotonForge"]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "photonforge>=1.3.0",
    "tidy3d>=2.10",
    "pyyaml>=6.0",
    "gdstk>=0.9",
    "numpy>=1.24",
]

[project.optional-dependencies]
test = ["pytest>=7.2"]

[project.urls]
Source = "https://github.com/flexcompute/cornerstone-forge"
"Cornerstone PDK" = "https://github.com/cornerstone-uos/cornerstone-pdk"
PhotonForge = "https://docs.flexcompute.com/projects/photonforge/"

[tool.setuptools.packages.find]
include = ["cornerstone_forge*"]

[tool.setuptools.package-data]
cornerstone_forge = [
    "library/*/*.yaml",
    "library/*/components/*.gds",
    "library/*/components/*.yaml",
    "library/*/cross-sections/*.yaml",
    "library/*/floorplans/*.yaml",
]
